+++ R/quadmesh.R
@@ -29,8 +29,8 @@ prs <- function(x) {
 
 #' Create a quad-type mesh for use in rgl.
 #'
-#' Convert an object to a \code{\link[rgl]{mesh3d}} quadrangle mesh,
-#' currently the only available method is for \code{\link[raster]{raster}}.
+#' Convert an object to a `mesh3d` ([rgl::qmesh3d()]) quadrangle mesh,
+#' with methods for [raster::raster()] and `matrix`.
 #'
 #' `quadmesh()` generates the cell-based interpretation of a raster (AREA) but applies a continuous
 #' interpretation of the values of the cells to each quad corner. `dquadmesh` splits the mesh and